example/requirements.txt
fastapi
uvicorn
aiohttp
slack-bolt
slack-click
